fix: some items to work properly
